Namespace canonical properties as Canonical::Property
