+++ app/models/property.rb
@@ -2,7 +2,7 @@
 
 class Property < ApplicationRecord
   belongs_to :game
-  belongs_to :canonical_property
+  belongs_to :canonical_property, class_name: 'Canonical::Property'
 
   has_many :shopping_lists, dependent: nil
   has_many :inventory_lists, dependent: nil
@@ -13,16 +13,16 @@ class Property < ApplicationRecord
 
   validates :name,
             presence:   true,
-            inclusion:  { in: CanonicalProperty::VALID_NAMES, message: "must be an ownable property in Skyrim, or the Arch-Mage's Quarters" },
+            inclusion:  { in: Canonical::Property::VALID_NAMES, message: "must be an ownable property in Skyrim, or the Arch-Mage's Quarters" },
             uniqueness: { scope: :game_id, message: 'must be unique per game' }
 
   validates :hold,
             presence:   true,
-            inclusion:  { in: CanonicalProperty::VALID_HOLDS, message: 'must be one of the nine Skyrim holds, or Solstheim' },
+            inclusion:  { in: Canonical::Property::VALID_HOLDS, message: 'must be one of the nine Skyrim holds, or Solstheim' },
             uniqueness: { scope: :game_id, message: 'must be unique per game' }
 
   validates :city,
-            inclusion:  { in: CanonicalProperty::VALID_CITIES, message: 'must be a Skyrim city in which an ownable property is located', allow_blank: true },
+            inclusion:  { in: Canonical::Property::VALID_CITIES, message: 'must be a Skyrim city in which an ownable property is located', allow_blank: true },
             uniqueness: { scope: :game_id, message: 'must be unique per game if present', allow_blank: true }
 
   validate :ensure_alchemy_lab_available,      if: -> { has_alchemy_lab == true && !canonical_property&.alchemy_lab_available }
@@ -33,12 +33,12 @@ class Property < ApplicationRecord
   private
 
   def ensure_max
-    Rails.logger.error "Cannot create property \"#{name}\" in hold \"#{hold}\": this game already has #{CanonicalProperty::TOTAL_PROPERTY_COUNT} properties"
+    Rails.logger.error "Cannot create property \"#{name}\" in hold \"#{hold}\": this game already has #{Canonical::Property::TOTAL_PROPERTY_COUNT} properties"
     errors.add(:game, 'already has max number of ownable properties')
   end
 
   def count_is_max
-    game.present? && game.properties.count == CanonicalProperty::TOTAL_PROPERTY_COUNT
+    game.present? && game.properties.count == Canonical::Property::TOTAL_PROPERTY_COUNT
   end
 
   def ensure_alchemy_lab_available
